requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
pydantic
pydantic-settings
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
schedule==1.2.0
alembic==1.13.0